Tidy up handling message processing spans, add span status
